Preencher produtos

+++ db_criacao/configura_banco_de_dados_rapido/criar/scripts/035-dimensoes.sql
@@ -1,7 +1,7 @@
 CREATE TABLE "dimensoes"(
     "id"                SERIAL PRIMARY KEY  NOT NULL,
-    "largura"           SMALLINT            NOT NULL,
-    "altura"            SMALLINT            NOT NULL,
-    "profundidade"      SMALLINT            NOT NULL,
-    "unidade_medida"    SMALLINT            NOT NULL
+    "largura"           INTEGER             NOT NULL,
+    "altura"            INTEGER             NOT NULL,
+    "profundidade"      INTEGER             NOT NULL,
+    "unidade_medida"    INTEGER             NOT NULL
 );
